Set up registry and API
